requirements.txt
django
pandas
numpy
statsmodels
requests
django-extensions
django-cors-headers
djangorestframework
